+++ src/Domain/Model/ChatMessageLinkableInterface.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Proximum\Vimeet\Domain\Model;
+
+interface ChatMessageLinkableInterface
+{
+    /**
+     * @return int
+     */
+    public function getId();
+
+    public function getObjectType(): string;
+
+    public function getEvent(): Event;
+}