+++ src/Ui/Bundle/EventBundle/Resources/assets/js/main.js
@@ -13,7 +13,7 @@ import EditableTextIndicator from './components/_EditableTextIndicator';
 import ProductSelector from './components/_ProductSelector';
 import QuantitySelector from './components/_QuantitySelector';
 import CatalogSheetCard from './components/_CatalogSheetCard';
-import {init as initAgenda} from './components/agenda';
+import { init as initAgenda } from './components/agenda';
 import AgendaRefresh from './components/agenda/_Refresh';
 import AgendaAllSheet from './components/agenda/_AgendaAllSheet';
 import Program from './components/agenda/_Program';
@@ -35,13 +35,16 @@ import SortParticipants from './components/_SortParticipants';
 import DateTimePicker from '../../../../../../../assets/js/components/DateTimePicker';
 import addSubmitEventListenerOnElementChange from './components/form/_AddSubmitEventListenerOnElementChange';
 import SheetVideo from './components/_SheetVideo';
+import UserConnectionRegister from './components/_UserConnectionRegister';
+import NotificationToast from './components/_NotificationToast';
+import initNetworking from './networking'
 
 import 'bootstrap';
 import 'elao-form.js';
 import 'intl-tel-input';
 import 'select2';
 
-function init (target) {
+function init(target) {
     // always first one in order to avoid collision
     [].forEach.call(target.querySelectorAll('.select2'), function (element) {
         $(element).select2({
@@ -67,7 +70,7 @@ function init (target) {
     });
 
     [].forEach.call(target.querySelectorAll('[data-company-info-update]'), function () {
-        var anchor         = window.location.hash.substring(1);
+        var anchor = window.location.hash.substring(1);
         var anchorElements = target.getElementsByName(anchor);
 
         if (anchor !== '' && anchor !== null && anchorElements.length > 0) {
@@ -132,8 +135,8 @@ function init (target) {
         new AgendaAllSheet(agendaAllSheetElement);
     }
 
-    [].forEach.call(target.querySelectorAll('.program-happening, .program-mass'), function(element) {
-       new Program(element);
+    [].forEach.call(target.querySelectorAll('.program-happening, .program-mass'), function (element) {
+        new Program(element);
     });
 
     [].forEach.call(target.querySelectorAll('.catalog__meeting_request'), function (element) {
@@ -158,8 +161,8 @@ function init (target) {
 
     $('#navigation-mobile, .mobile-menu .navigation__close', target).on('click', function (e) {
         $('.mobile-menu').toggle();
-        setTimeout(function() {
-          $('body').toggleClass('menu-mobile-opened').scrollTop(0);
+        setTimeout(function () {
+            $('body').toggleClass('menu-mobile-opened').scrollTop(0);
         }, 1);
     });
 
@@ -175,7 +178,7 @@ function init (target) {
         .on('loaded.bs.modal', function (event) {
             PubSub.publish('dom.added', event.target);
         }.bind(this))
-    ;
+        ;
 
     $('.show-modal', target).modal('show');
 
@@ -278,6 +281,14 @@ function init (target) {
     [].forEach.call(target.querySelectorAll('.sort-participants'), function (element) {
         new SortParticipants(element);
     });
+
+    const userConnectionRegister = new UserConnectionRegister();
+    userConnectionRegister.connect();
+
+    initNetworking(document, userConnectionRegister);
+
+    new NotificationToast(target.querySelector('[data-notification-toast]'), userConnectionRegister);
+
 }
 
 PubSub.subscribe('dom.added', function (name, element) { init(element); });