+++ tests/Application/Command/Happening/Webinar/Question/VoteHappeningQuestionHandlerTest.php
@@ -4,10 +4,12 @@ namespace Application\Command\Happening\Webinar\Question;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
+use Proximum\Vimeet\Application\Adapter\NotificationPublisherInterface;
 use Proximum\Vimeet\Application\Command\Happening\Webinar\Question\VoteHappeningQuestion;
 use Proximum\Vimeet\Application\Command\Happening\Webinar\Question\VoteHappeningQuestionHandler;
 use Proximum\Vimeet\Application\Exception\Happening\QuestionNotAllowedException;
 use Proximum\Vimeet\Application\Exception\Happening\QuestionNotFoundException;
+use Proximum\Vimeet\Domain\Model\Happening;
 use Proximum\Vimeet\Domain\Model\Happening\Question;
 use Proximum\Vimeet\Domain\Model\Happening\QuestionVote;
 use Proximum\Vimeet\Domain\Model\User;
@@ -22,6 +24,9 @@ class VoteHappeningQuestionHandlerTest extends TestCase
     /** @var ObjectProphecy */
     private $questionVoteRepository;
 
+    /** @var ObjectProphecy */
+    private $notificationPublisher;
+
     /** @var VoteHappeningQuestionHandler */
     private $voteHappeningQuestionHandler;
 
@@ -29,7 +34,13 @@ class VoteHappeningQuestionHandlerTest extends TestCase
     {
         $this->questionRepository = $this->prophesize(QuestionRepositoryInterface::class);
         $this->questionVoteRepository = $this->prophesize(QuestionVoteRepositoryInterface::class);
-        $this->voteHappeningQuestionHandler = new VoteHappeningQuestionHandler($this->questionRepository->reveal(), $this->questionVoteRepository->reveal());
+        $this->notificationPublisher = $this->prophesize(NotificationPublisherInterface::class);
+
+        $this->voteHappeningQuestionHandler = new VoteHappeningQuestionHandler(
+            $this->questionRepository->reveal(),
+            $this->questionVoteRepository->reveal(),
+            $this->notificationPublisher->reveal()
+        );
     }
 
     public function test_vote_happening_question()
@@ -49,6 +60,7 @@ class VoteHappeningQuestionHandlerTest extends TestCase
         $questionAuthor->getId()->shouldBeCalled()->willReturn(24);
         $question = $this->prophesize(Question::class);
         $question->getCreatedBy()->shouldBeCalled()->willReturn($questionAuthor->reveal());
+        $question->getHappening()->shouldBeCalled()->willReturn($this->prophesize(Happening::class));
 
         $this->questionRepository->findById(42)->shouldBeCalled()->willReturn($question->reveal());
         $this->questionVoteRepository
@@ -78,6 +90,7 @@ class VoteHappeningQuestionHandlerTest extends TestCase
         $questionAuthor->getId()->shouldBeCalled()->willReturn(24);
         $question = $this->prophesize(Question::class);
         $question->getCreatedBy()->shouldBeCalled()->willReturn($questionAuthor->reveal());
+        $question->getHappening()->shouldBeCalled()->willReturn($this->prophesize(Happening::class));
 
         $this->questionRepository->findById(42)->shouldBeCalled()->willReturn($question->reveal());